\name{gpu_session_backend}
\alias{gpu_session_backend}
\title{Return the Backend Associated with a Session}
\usage{
gpu_session_backend(session)
}
\arguments{
\item{session}{A session handle created with \code{gpu_session_open()}.}
}
\value{
Backend name.
}
\description{
Returns the backend associated with a session handle.
}
